+++ source/DSCResources/DSC_SqlSetup/DSC_SqlSetup.psm1
@@ -176,6 +176,7 @@ function Get-TargetResource
         UseEnglish                 = $UseEnglish
         ServerName                 = $ServerName
         SqlVersion                 = $null
+        ProductCoveredBySA         = $null
     }
 
     <#
@@ -316,6 +317,16 @@ function Get-TargetResource
             $getTargetResourceReturnValue.SqlTempdbFileGrowth = $currentTempDbProperties.SqlTempdbFileGrowth
             $getTargetResourceReturnValue.SqlTempdbLogFileSize = $currentTempDbProperties.SqlTempdbLogFileSize
             $getTargetResourceReturnValue.SqlTempdbLogFileGrowth = $currentTempDbProperties.SqlTempdbLogFileGrowth
+        }
+
+        if ($sqlVersion -ge 16)
+        {
+            # Grab the value of ProductCoveredBySA from the registry based on the instance
+            $getRegistryPropertyParams = @{
+                Path = "HKLM:\SOFTWARE\Microsoft\Microsoft SQL Server\MSSQL$($SqlVersion).$($InstanceName)\Setup"
+                Name = 'IsProductCoveredBySA'
+        }
+            $getTargetResourceReturnValue.ProductCoveredBySA = Get-RegistryPropertyValue @getRegistryPropertyParams
         }
 
         # Get all members of the sysadmin role.
@@ -517,6 +528,11 @@ function Get-TargetResource
 
     .PARAMETER ProductKey
         Product key for licensed installations.
+
+    .PARAMETER ProductCoveredBySA
+        Specifies the license coverage for SQL Server. True indicates it's covered under Software Assurance or SQL Server subscription.
+        False, or omitting the parameter, indicates it's covered under a SQL Server license.
+        Default value is False.
 
     .PARAMETER UpdateEnabled
         Enabled updates during installation.
@@ -747,6 +763,10 @@ function Set-TargetResource
         [Parameter()]
         [System.String]
         $ProductKey,
+
+        [Parameter()]
+        [System.Boolean]
+        $ProductCoveredBySA,
 
         [Parameter()]
         [System.String]
@@ -1290,6 +1310,12 @@ function Set-TargetResource
         $setupArguments['FailoverClusterIPAddresses'] = $clusterIPAddresses
     }
 
+    # Add Parameter ProductCoveredBySA
+    if ($PSBoundParameters.ContainsKey('ProductCoveredBySA'))
+    {
+        $setupArguments['ProductCoveredBySA'] = $ProductCoveredBySA
+    }
+
     # Add standard install arguments
     $setupArguments += @{
         Quiet                        = $true
@@ -1767,6 +1793,13 @@ function Set-TargetResource
     .PARAMETER ProductKey
         Product key for licensed installations.
 
+    .PARAMETER ProductCoveredBySA
+        Specifies the license coverage for SQL Server. True indicates it's covered under Software Assurance or SQL Server subscription.
+        False, or omitting the parameter, indicates it's covered under a SQL Server license.
+        Default value is False.
+
+        Not used in Test-TargetResource.    
+
     .PARAMETER UpdateEnabled
         Enabled updates during installation.
 
@@ -2004,6 +2037,10 @@ function Test-TargetResource
         [Parameter()]
         [System.String]
         $ProductKey,
+
+        [Parameter()]
+        [System.Boolean]
+        $ProductCoveredBySA,
 
         [Parameter()]
         [System.String]