+++ tests/Unit/DSC_SqlSetup.Tests.ps1
@@ -33,6 +33,9 @@ BeforeDiscovery {
     # Testing each supported SQL Server version
     $testProductVersion = @(
         @{
+            MockSqlMajorVersion = 16 # SQL Server 2022
+        }
+        @{
             MockSqlMajorVersion = 15 # SQL Server 2019
         }
         @{
@@ -332,11 +335,11 @@ Describe 'SqlSetup\Get-TargetResource' -Tag 'Get' {
 
             InModuleScope -ScriptBlock {
                 $script:mockGetTargetResourceParameters = @{
-                    InstanceName     = 'MSSQLSERVER'
-                    SourceCredential = $null
-                    SourcePath       = $TestDrive
-                    Feature          = 'NewFeature' # Test enabling a code-feature.
-                    ServerName       = 'host.company.local'
+                    InstanceName       = 'MSSQLSERVER'
+                    SourceCredential   = $null
+                    SourcePath         = $TestDrive
+                    Feature            = 'NewFeature' # Test enabling a code-feature.
+                    ServerName         = 'host.company.local'
                 }
             }
         }
@@ -394,6 +397,7 @@ Describe 'SqlSetup\Get-TargetResource' -Tag 'Get' {
                 $result.RSSvcAccountUsername | Should -BeNullOrEmpty
                 $result.ASSvcAccountUsername | Should -BeNullOrEmpty
                 $result.ASCollation | Should -BeNullOrEmpty
+                $result.ProductCoveredBySA | Should -BeFalse
                 $result.ASSysAdminAccounts | Should -BeNullOrEmpty
                 $result.ASDataDir | Should -BeNullOrEmpty
                 $result.ASLogDir | Should -BeNullOrEmpty
@@ -658,10 +662,10 @@ Describe 'SqlSetup\Get-TargetResource' -Tag 'Get' {
                 $script:mockSourcePathUNC = Join-Path -Path "\\localhost\$testDrive_DriveShare" -ChildPath (Split-Path -Path $TestDrive -NoQualifier)
 
                 $script:mockGetTargetResourceParameters = @{
-                    InstanceName     = 'MSSQLSERVER'
-                    SourceCredential = New-Object -TypeName System.Management.Automation.PSCredential -ArgumentList @('COMPANY\sqladmin', ('dummyPassw0rd' | ConvertTo-SecureString -asPlainText -Force))
-                    SourcePath       = $mockSourcePathUNC
-                    SqlVersion       = ('{0}.0' -f $MockSqlMajorVersion)
+                    InstanceName       = 'MSSQLSERVER'
+                    SourceCredential   = New-Object -TypeName System.Management.Automation.PSCredential -ArgumentList @('COMPANY\sqladmin', ('dummyPassw0rd' | ConvertTo-SecureString -asPlainText -Force))
+                    SourcePath         = $mockSourcePathUNC
+                    SqlVersion         = ('{0}.0' -f $MockSqlMajorVersion)
                 }
             }
         }
@@ -828,6 +832,18 @@ Describe 'SqlSetup\Get-TargetResource' -Tag 'Get' {
                     $result.Features | Should -Match 'BOL\b'
                     $result.Features | Should -Match 'MDS\b'
                 }
+                elseif ($MockSqlMajorVersion -in ('16'))
+                {
+                    $result.Features | Should -Match 'SQLENGINE\b'
+                    $result.Features | Should -Match 'REPLICATION\b'
+                    $result.Features | Should -Match 'DQ\b'
+                    $result.Features | Should -Match 'DQC\b'
+                    $result.Features | Should -Match 'FULLTEXT\b'
+                    $result.Features | Should -Match 'AS\b'
+                    $result.Features | Should -Match 'IS\b'
+                    $result.Features | Should -Match 'BOL\b'
+                    $result.Features | Should -Match 'MDS\b'
+                }
                 else
                 {
                     $result.Features | Should -Match 'SQLENGINE\b'
@@ -971,6 +987,18 @@ Describe 'SqlSetup\Get-TargetResource' -Tag 'Get' {
                     $result.Features | Should -Match 'CONN\b'
                     $result.Features | Should -Match 'BC\b'
                     $result.Features | Should -Match 'SDK\b'
+                }
+                elseif ($MockSqlMajorVersion -in ('16'))
+                {
+                    $result.Features | Should -Match 'SQLENGINE\b'
+                    $result.Features | Should -Match 'REPLICATION\b'
+                    $result.Features | Should -Match 'DQ\b'
+                    $result.Features | Should -Match 'DQC\b'
+                    $result.Features | Should -Match 'FULLTEXT\b'
+                    $result.Features | Should -Match 'AS\b'
+                    $result.Features | Should -Match 'IS\b'
+                    $result.Features | Should -Match 'BOL\b'
+                    $result.Features | Should -Match 'MDS\b'
                 }
                 else
                 {
@@ -1204,6 +1232,18 @@ Describe 'SqlSetup\Get-TargetResource' -Tag 'Get' {
                     $result.Features | Should -Match 'CONN\b'
                     $result.Features | Should -Match 'BC\b'
                     $result.Features | Should -Match 'SDK\b'
+                }
+                elseif ($MockSqlMajorVersion -in ('16'))
+                {
+                    $result.Features | Should -Match 'SQLENGINE\b'
+                    $result.Features | Should -Match 'REPLICATION\b'
+                    $result.Features | Should -Match 'DQ\b'
+                    $result.Features | Should -Match 'DQC\b'
+                    $result.Features | Should -Match 'FULLTEXT\b'
+                    $result.Features | Should -Match 'AS\b'
+                    $result.Features | Should -Match 'IS\b'
+                    $result.Features | Should -Match 'BOL\b'
+                    $result.Features | Should -Match 'MDS\b'
                 }
                 else
                 {
@@ -1513,6 +1553,18 @@ Describe 'SqlSetup\Get-TargetResource' -Tag 'Get' {
                     $result.Features | Should -Match 'CONN\b'
                     $result.Features | Should -Match 'BC\b'
                     $result.Features | Should -Match 'SDK\b'
+                }
+                elseif ($MockSqlMajorVersion -in ('16'))
+                {
+                    $result.Features | Should -Match 'SQLENGINE\b'
+                    $result.Features | Should -Match 'REPLICATION\b'
+                    $result.Features | Should -Match 'DQ\b'
+                    $result.Features | Should -Match 'DQC\b'
+                    $result.Features | Should -Match 'FULLTEXT\b'
+                    $result.Features | Should -Match 'AS\b'
+                    $result.Features | Should -Match 'IS\b'
+                    $result.Features | Should -Match 'BOL\b'
+                    $result.Features | Should -Match 'MDS\b'
                 }
                 else
                 {
@@ -2315,6 +2367,10 @@ Describe 'SqlSetup\Set-TargetResource' -Tag 'Set' {
                     {
                         $mockStartSqlSetupProcessExpectedArgument.Features = 'SQLENGINE,REPLICATION,DQ,DQC,FULLTEXT,AS,IS,BOL,CONN,BC,SDK,MDS'
                     }
+                    elseif ($MockSqlMajorVersion -in ('16'))
+                    {
+                        $mockStartSqlSetupProcessExpectedArgument.Features = 'SQLENGINE,REPLICATION,DQ,DQC,FULLTEXT,AS,IS,BOL,MDS'
+                    }
                     else
                     {
                         $mockStartSqlSetupProcessExpectedArgument.Features = 'SQLENGINE,REPLICATION,DQ,DQC,FULLTEXT,RS,AS,IS,BOL,CONN,BC,SDK,MDS,SSMS,ADV_SSMS'
@@ -2351,7 +2407,7 @@ Describe 'SqlSetup\Set-TargetResource' -Tag 'Set' {
                             BrowserSvcStartupType  = 'Automatic'
                         }
 
-                        if ($MockSqlMajorVersion -in ('13', '14', '15'))
+                        if ($MockSqlMajorVersion -in ('13', '14', '15', '16'))
                         {
                             $mockSetTargetResourceParameters.Features = $mockSetTargetResourceParameters.Features -replace ',SSMS,ADV_SSMS', ''
                         }
@@ -2364,6 +2420,13 @@ Describe 'SqlSetup\Set-TargetResource' -Tag 'Set' {
                         elseif ($MockSqlMajorVersion -in ('14', '15'))
                         {
                             $mockSetTargetResourceParameters.Features = $mockSetTargetResourceParameters.Features -replace ',RS', ''
+                        }
+                        elseif ($MockSqlMajorVersion -in ('16'))
+                        {
+                            $mockSetTargetResourceParameters.Features = $mockSetTargetResourceParameters.Features -replace ',Conn', ''
+                            $mockSetTargetResourceParameters.Features = $mockSetTargetResourceParameters.Features -replace ',RS', ''
+                            $mockSetTargetResourceParameters.Features = $mockSetTargetResourceParameters.Features -replace ',Bc', ''
+                            $mockSetTargetResourceParameters.Features = $mockSetTargetResourceParameters.Features -replace ',SDK', ''
                         }
 
                         { Set-TargetResource @mockSetTargetResourceParameters } | Should -Not -Throw
@@ -2598,6 +2661,10 @@ Describe 'SqlSetup\Set-TargetResource' -Tag 'Set' {
                     {
                         $mockStartSqlSetupProcessExpectedArgument.Features = 'SQLENGINE,REPLICATION,DQ,DQC,FULLTEXT,AS,IS,BOL,CONN,BC,SDK,MDS'
                     }
+                    elseif ($MockSqlMajorVersion -in ('16'))
+                    {
+                        $mockStartSqlSetupProcessExpectedArgument.Features = 'SQLENGINE,REPLICATION,DQ,DQC,FULLTEXT,AS,IS,BOL,MDS'
+                    }
                     else
                     {
                         $mockStartSqlSetupProcessExpectedArgument.Features = 'SQLENGINE,REPLICATION,DQ,DQC,FULLTEXT,RS,AS,IS,BOL,CONN,BC,SDK,MDS,SSMS,ADV_SSMS'
@@ -2636,7 +2703,7 @@ Describe 'SqlSetup\Set-TargetResource' -Tag 'Set' {
                             SqlVersion             = ('{0}.0' -f $MockSqlMajorVersion)
                         }
 
-                        if ($MockSqlMajorVersion -in ('13', '14', '15'))
+                        if ($MockSqlMajorVersion -in ('13', '14', '15', '16'))
                         {
                             $mockSetTargetResourceParameters.Features = $mockSetTargetResourceParameters.Features -replace ',SSMS,ADV_SSMS', ''
                         }
@@ -2649,6 +2716,13 @@ Describe 'SqlSetup\Set-TargetResource' -Tag 'Set' {
                         elseif ($MockSqlMajorVersion -in ('14', '15'))
                         {
                             $mockSetTargetResourceParameters.Features = $mockSetTargetResourceParameters.Features -replace ',RS', ''
+                        }
+                        elseif ($MockSqlMajorVersion -in ('16'))
+                        {
+                            $mockSetTargetResourceParameters.Features = $mockSetTargetResourceParameters.Features -replace ',Conn', ''
+                            $mockSetTargetResourceParameters.Features = $mockSetTargetResourceParameters.Features -replace ',RS', ''
+                            $mockSetTargetResourceParameters.Features = $mockSetTargetResourceParameters.Features -replace ',Bc', ''
+                            $mockSetTargetResourceParameters.Features = $mockSetTargetResourceParameters.Features -replace ',SDK', ''
                         }
 
                         { Set-TargetResource @mockSetTargetResourceParameters } | Should -Not -Throw
@@ -2706,6 +2780,52 @@ Describe 'SqlSetup\Set-TargetResource' -Tag 'Set' {
                 Should -Invoke -CommandName Start-SqlSetupProcess -Exactly -Times 1 -Scope It
             }
         }
+
+        Context 'When installing the database engine and ProductcoveredBySA is true' {
+                BeforeAll {
+                    Mock -CommandName Get-FilePathMajorVersion -MockWith {
+                        return 16
+                    }
+
+                    Mock -CommandName Get-TargetResource -MockWith {
+                        return @{
+                            Features = ''
+                        }
+                    }
+                }
+
+                It 'Should set the system in the desired state when feature is SQLENGINE' {
+                    $mockStartSqlSetupProcessExpectedArgument = @{
+                        Quiet                        = 'True'
+                        IAcceptSQLServerLicenseTerms = 'True'
+                        Action                       = 'Install'
+                        InstanceName                 = 'MSSQLSERVER'
+                        Features                     = 'SQLENGINE'
+                        SQLSysAdminAccounts          = 'COMPANY\sqladmin COMPANY\SQLAdmins COMPANY\User1'
+                        PID                          = '1FAKE-2FAKE-3FAKE-4FAKE-5FAKE'
+                        ProductCoveredBySA           = 'True'
+                    }
+
+                    InModuleScope -ScriptBlock {
+                        Set-StrictMode -Version 1.0
+
+                        $mockSetTargetResourceParameters = @{
+                            Features            = 'SQLENGINE'
+                            SQLSysAdminAccounts = 'COMPANY\User1', 'COMPANY\SQLAdmins'
+                            InstanceName        = 'MSSQLSERVER'
+                            SourceCredential    = $null
+                            SourcePath          = $TestDrive
+                            ProductKey          = '1FAKE-2FAKE-3FAKE-4FAKE-5FAKE'
+                            ProductCoveredBySA  = $true
+                            ServerName          = 'host.company.local'
+                        }
+
+                        { Set-TargetResource @mockSetTargetResourceParameters } | Should -Not -Throw
+                    }
+
+                    Should -Invoke -CommandName Start-SqlSetupProcess -Exactly -Times 1 -Scope It
+                }
+            }
 
         Context 'When installing the database engine and disabling the TCP protocol' {
             BeforeAll {
@@ -2964,6 +3084,9 @@ Describe 'SqlSetup\Set-TargetResource' -Tag 'Set' {
     }
 
     Context 'When passing invalid features for <MockSqlMajorVersion>' -ForEach @(
+        @{
+            MockSqlMajorVersion = 16 # SQL Server 2022
+        }
         @{
             MockSqlMajorVersion = 15 # SQL Server 2019
         }